requirements.txt
absl-py==1.3.0
anndata==0.8.0
anyio==3.6.2
appnope==0.1.3
arch==5.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asteval==0.9.28
astunparse==1.6.3
async-lru==2.0.4
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==2.2.0
contourpy==1.0.6
cospar==0.3.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
EMD-signal==1.6.0
EntropyHub==2.0
entrypoints==0.4
ete3==3.1.2
eval_type_backport==0.2.0
ewstools==2.1.2
fastcluster==1.2.6
fastjsonschema==2.16.3
flatbuffers==22.10.26
fonttools==4.38.0
fqdn==1.5.1
future==0.18.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1653138941901/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
idna==3.4
igraph==0.10.2
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==7.4.7
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.24.0
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
leidenalg==0.9.0
libclang==14.0.6
llvmlite==0.39.1
lmfit==1.0.3
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mizani==0.8.1
multiprocess==0.70.16
natsort==8.2.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1660672958449/work
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
palettable==3.3.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.16.1
plotnine==0.10.1
pox==0.3.4
ppft==1.7.6.8
prometheus-client==0.16.0
prompt-toolkit==3.0.33
property-cached==1.6.4
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
scanpy==1.9.1
scikit-learn==1.1.3
scikit-misc==0.1.4
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==2.2.0
session-info==1.0.0
setproctitle==1.3.3
shtab==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
spyder-kernels==2.3.0
statsmodels==0.13.2
stdlib-list==0.8.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
termcolor==2.1.1
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.11.0
tyro==0.8.4
umap-learn==0.5.3
uncertainties==3.1.7
uri-template==1.2.0
urllib3==1.26.12
wandb==0.17.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wurlitzer==3.0.2
zipp==3.10.0
